Bug fixing
Added in a function to calculate line strain
changed the formula for youngs modulus so that it uses m+c as the max value.
Created_balanced_blocks: the power in two node blocks is equal to the power in each node
Find_network_balance: can handle blocks with no net power.